+++ map_mapper/map_mapper_annotations/lib/config.g.dart
@@ -19,8 +19,8 @@ Config _$ConfigFromJson(Map<String, dynamic> json) => $checkedCreate(
           ],
         );
         final val = Config(
-          keyHandlerType: $checkedConvert(
-              'keyHandlerType', (v) => v as String? ?? 'DefaultKeyHandler'),
+          keyHandlerType:
+              $checkedConvert('keyHandlerType', (v) => v as String? ?? ''),
           durationPrecision: $checkedConvert(
               'durationPrecision',
               (v) =>